requirements.txt
absl-py==0.1.11
bleach==1.5.0
Click==7.0
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
Flask==1.0.2
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.8.2
ipywidgets==7.1.2
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
numpy==1.14.2
opencv-python==3.4.3.18
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.5.1
tensorflow-tensorboard==1.5.1
terminado==0.8.1
testpath==0.3.1
tornado==5.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4